chore: split local_model_server into _section files via include!
