+++ src/local_model_server/mod.rs
@@ -21,1590 +21,14 @@ use sha2::{Digest, Sha256};
 use crate::llm::{EmbeddingBackend, EmbeddingInputKind};
 use crate::local_backend::{LocalProgressReporter, default_local_model_cache_dir};
 
-pub(crate) const MODEL_SERVER: &[u8] = include_bytes!(concat!(
-    env!("CARGO_MANIFEST_DIR"),
-    "/components/model_server/rara_model_server.py"
-));
-pub(crate) const MODEL_SERVER_NAME: &str = "rara_model_server.py";
-
-pub(crate) const REQUIREMENTS_MACOS_ARM64: &[u8] = include_bytes!(concat!(
-    env!("CARGO_MANIFEST_DIR"),
-    "/components/model_server/requirements-macos-arm64.txt"
-));
-pub(crate) const REQUIREMENTS_MACOS_ARM64_NAME: &str = "requirements-macos-arm64.txt";
-
-pub(crate) const REQUIREMENTS_PORTABLE: &[u8] = include_bytes!(concat!(
-    env!("CARGO_MANIFEST_DIR"),
-    "/components/model_server/requirements-portable.txt"
-));
-pub(crate) const REQUIREMENTS_PORTABLE_NAME: &str = "requirements-portable.txt";
-pub(crate) const MLX_QWEN3_MODEL_ID: &str = "mlx-community/Qwen3-Embedding-0.6B-4bit-DWQ";
-pub(crate) const FASTEMBED_BGE_M3_MODEL_ID: &str = "BAAI/bge-m3";
-pub(crate) const DEFAULT_MODEL_SERVER_HOST: &str = "127.0.0.1";
-pub(crate) const DEFAULT_MODEL_SERVER_PORT: u16 = 18181;
-pub(crate) const SERVER_METADATA_NAME: &str = "server.json";
-pub(crate) const STARTUP_LOCK_NAME: &str = "startup.lock";
-pub(crate) const REQUIREMENTS_MARKER_NAME: &str = "requirements-installed.json";
-pub(crate) const MODEL_SNAPSHOT_MARKER_NAME: &str = "model-snapshot.json";
-pub(crate) const HEALTH_TIMEOUT: Duration = Duration::from_millis(300);
-pub(crate) const PREPARE_TIMEOUT: Duration = Duration::from_secs(30);
-pub(crate) const STARTUP_HEALTH_ATTEMPTS: usize = 30;
-pub(crate) const STARTUP_HEALTH_DELAY: Duration = Duration::from_millis(200);
-pub(crate) const MODEL_REVISION: &str = "main";
-pub(crate) static ATOMIC_WRITE_COUNTER: AtomicU64 = AtomicU64::new(0);
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum SnapshotPreparation {
-    RustManaged {
-        required_files: SnapshotRequiredFiles,
-    },
-    PythonManaged,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum SnapshotRequiredFiles {
-    MlxQwen3,
-}
-
-#[derive(Debug, Clone, Copy)]
-pub(crate) struct LocalEmbeddingModelProfile {
-    backend: &'static str,
-    model: &'static str,
-    revision: &'static str,
-    snapshot_preparation: SnapshotPreparation,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum LocalModelServerState {
-    Ready,
-    Starting,
-    WaitingForServer,
-    CreatingVenv,
-    InstallingDependencies,
-    PreparingModel,
-    PreparedButStopped,
-    SetupRequired,
-    Error,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct LocalModelServerStatus {
-    pub state: LocalModelServerState,
-    pub backend: String,
-    pub model: String,
-    pub detail: String,
-    pub server_path: Option<PathBuf>,
-    pub endpoint: Option<String>,
-}
-
-impl Default for LocalModelServerStatus {
-    fn default() -> Self {
-        let (backend, model) = default_embedding_backend();
-        Self {
-            state: LocalModelServerState::SetupRequired,
-            backend: backend.to_string(),
-            model: model.to_string(),
-            detail: "model server component has not been prepared".to_string(),
-            server_path: None,
-            endpoint: None,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct BundledModelServer {
-    pub path: PathBuf,
-    pub sha256: String,
-    pub runtime_dir: PathBuf,
-    pub venv_dir: PathBuf,
-    pub requirements: Vec<BundledModelServerFile>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct BundledModelServerFile {
-    pub path: PathBuf,
-    pub sha256: String,
-}
-
-pub(crate) struct BundledFile {
-    name: &'static str,
-    content: &'static [u8],
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(crate) struct ModelServerMetadata {
-    pid: u32,
-    host: String,
-    port: u16,
-    component_sha256: String,
-    profile: String,
-    started_at: u64,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(crate) struct ModelSnapshotMarker {
-    model: String,
-    revision: String,
-    snapshot_path: PathBuf,
-    files: Vec<String>,
-}
-
-pub(crate) struct StartupLock {
-    _file: File,
-}
-
-pub(crate) struct LocalModelServerEmbeddingBackend {
-    rara_home: PathBuf,
-    client: reqwest::Client,
-    endpoint: Mutex<Option<String>>,
-}
-
-impl LocalModelServerEmbeddingBackend {
-    pub(crate) fn new(rara_home: PathBuf) -> Result<Self> {
-        let status = prepare_local_model_server_status(&rara_home);
-        Self::from_initial_status(rara_home, status)
-    }
-
-    pub(crate) fn from_initial_status(
-        rara_home: PathBuf,
-        status: LocalModelServerStatus,
-    ) -> Result<Self> {
-        Ok(Self {
-            rara_home,
-            client: reqwest::Client::builder()
-                .timeout(Duration::from_secs(30))
-                .no_proxy()
-                .build()
-                .context("build local embedding HTTP client")?,
-            endpoint: Mutex::new(status.endpoint),
-        })
-    }
-
-    async fn embed_once(
-        &self,
-        endpoint: &str,
-        text: &str,
-        kind: EmbeddingInputKind,
-    ) -> Result<Vec<f32>> {
-        let embeddings_url = format!("{}/v1/embeddings", endpoint.trim_end_matches('/'));
-        let response = self
-            .client
-            .post(&embeddings_url)
-            .json(&serde_json::json!({
-                "input": text,
-                "input_type": kind.as_api_value(),
-            }))
-            .send()
-            .await
-            .with_context(|| {
-                format!(
-                    "send local embedding request to {}",
-                    sanitize_url_for_display(&embeddings_url)
-                )
-            })?;
-        if !response.status().is_success() {
-            let body = redact_secrets(response.text().await.unwrap_or_default());
-            bail!(
-                "local embedding request failed at {}: {}",
-                sanitize_url_for_display(&embeddings_url),
-                body
-            );
-        }
-        let payload: Value = response
-            .json()
-            .await
-            .context("parse local embedding response")?;
-        let embedding = payload
-            .get("data")
-            .and_then(Value::as_array)
-            .and_then(|items| items.first())
-            .and_then(|item| item.get("embedding"))
-            .and_then(Value::as_array)
-            .ok_or_else(|| anyhow!("local embedding response missing data[0].embedding"))?;
-        embedding
-            .iter()
-            .map(|value| {
-                value
-                    .as_f64()
-                    .map(|number| number as f32)
-                    .ok_or_else(|| anyhow!("local embedding vector contained non-numeric value"))
-            })
-            .collect()
-    }
-
-    fn cached_endpoint(&self) -> Option<String> {
-        self.endpoint
-            .lock()
-            .expect("local embedding endpoint lock poisoned")
-            .clone()
-    }
-
-    async fn refresh_endpoint(&self) -> Result<String> {
-        let rara_home = self.rara_home.clone();
-        let result = tokio::task::spawn_blocking(move || {
-            let status = inspect_local_model_server_status(&rara_home);
-            let Some(endpoint) = status.endpoint else {
-                return Err(anyhow!(
-                    "local embedding model server unavailable: state={:?}; {}",
-                    status.state,
-                    status.detail
-                ));
-            };
-            Ok(endpoint)
-        })
-        .await
-        .context("join local embedding endpoint refresh")?;
-
-        match result {
-            Ok(endpoint) => {
-                *self
-                    .endpoint
-                    .lock()
-                    .expect("local embedding endpoint lock poisoned") = Some(endpoint.clone());
-                Ok(endpoint)
-            }
-            Err(err) => {
-                *self
-                    .endpoint
-                    .lock()
-                    .expect("local embedding endpoint lock poisoned") = None;
-                Err(err)
-            }
-        }
-    }
-}
-
-impl Drop for LocalModelServerEmbeddingBackend {
-    fn drop(&mut self) {
-        let runtime_dir = self.rara_home.join("runtime").join("model-server");
-        if let Ok(Some(metadata)) = read_server_metadata(&metadata_path(&runtime_dir)) {
-            let _ = nix::sys::signal::kill(
-                Pid::from_raw(metadata.pid as i32),
-                nix::sys::signal::Signal::SIGTERM,
-            );
-        }
-    }
-}
-
-#[async_trait]
-impl EmbeddingBackend for LocalModelServerEmbeddingBackend {
-    async fn embed(&self, text: &str, kind: EmbeddingInputKind) -> Result<Vec<f32>> {
-        let endpoint = match self.cached_endpoint() {
-            Some(endpoint) => endpoint,
-            None => self.refresh_endpoint().await?,
-        };
-        match self.embed_once(&endpoint, text, kind).await {
-            Ok(vector) => Ok(vector),
-            Err(first_error) => {
-                let refreshed_endpoint = self.refresh_endpoint().await?;
-                if refreshed_endpoint == endpoint {
-                    return Err(first_error);
-                }
-                self.embed_once(&refreshed_endpoint, text, kind).await
-            }
-        }
-    }
-}
-
-pub(crate) fn ensure_bundled_model_server(rara_home: &Path) -> Result<BundledModelServer> {
-    let runtime_dir = rara_home.join("runtime").join("model-server");
-    fs::create_dir_all(&runtime_dir)
-        .with_context(|| format!("create {}", runtime_dir.display()))?;
-    ensure_runtime_dir_inside_home(rara_home, &runtime_dir)?;
-
-    let path = runtime_dir.join(MODEL_SERVER_NAME);
-    let expected_hash = sha256_hex(MODEL_SERVER);
-    if existing_file_matches(&path, &expected_hash)? {
-        return Ok(BundledModelServer {
-            path,
-            sha256: expected_hash,
-            runtime_dir: runtime_dir.clone(),
-            venv_dir: runtime_dir.join("venv"),
-            requirements: ensure_model_server_requirements(&runtime_dir)?,
-        });
-    }
-
-    write_file_atomically(&path, MODEL_SERVER)?;
-    Ok(BundledModelServer {
-        path,
-        sha256: expected_hash,
-        runtime_dir: runtime_dir.clone(),
-        venv_dir: runtime_dir.join("venv"),
-        requirements: ensure_model_server_requirements(&runtime_dir)?,
-    })
-}
-
-pub(crate) fn prepare_local_model_server_status(rara_home: &Path) -> LocalModelServerStatus {
-    prepare_local_model_server_status_with_progress(rara_home, None)
-}
-
-pub(crate) fn prepare_local_model_server_status_with_progress(
-    rara_home: &Path,
-    progress: Option<LocalProgressReporter>,
-) -> LocalModelServerStatus {
-    prepare_local_model_server_status_inner(rara_home, BootstrapMode::Automatic, progress)
-}
-
-pub(crate) fn inspect_local_model_server_status(rara_home: &Path) -> LocalModelServerStatus {
-    prepare_local_model_server_status_inner(rara_home, BootstrapMode::InspectOnly, None)
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum BootstrapMode {
-    Automatic,
-    InspectOnly,
-}
-
-pub(crate) fn format_error_chain(err: &anyhow::Error) -> String {
-    let mut chain = err.chain();
-    let mut out = match chain.next() {
-        Some(first) => first.to_string(),
-        None => return String::new(),
-    };
-    for cause in chain {
-        out.push_str("\nCaused by: ");
-        out.push_str(&cause.to_string());
-    }
-    out
-}
-
-pub(crate) fn prepare_local_model_server_status_inner(
-    rara_home: &Path,
-    mode: BootstrapMode,
-    progress: Option<LocalProgressReporter>,
-) -> LocalModelServerStatus {
-    let profile = default_embedding_profile();
-    let backend = profile.backend;
-    let model = profile.model;
-    match ensure_bundled_model_server(rara_home) {
-        Ok(server) => {
-            if let Some(status) = reusable_server_status(&server, backend, model) {
-                return status;
-            }
-
-            if mode == BootstrapMode::InspectOnly {
-                return inspect_prepared_runtime_status(&server, backend, model);
-            }
-
-            let lock_path = startup_lock_path(&server.runtime_dir);
-            let Ok(_lock) = StartupLock::try_acquire(&lock_path) else {
-                return LocalModelServerStatus {
-                    state: LocalModelServerState::WaitingForServer,
-                    backend: backend.to_string(),
-                    model: model.to_string(),
-                    detail: "waiting for another RARA process to finish model server startup"
-                        .to_string(),
-                    server_path: Some(server.path),
-                    endpoint: None,
-                };
-            };
-
-            if let Some(status) = reusable_server_status(&server, backend, model) {
-                return status;
-            }
-
-            let python = match prepared_runtime_python(&server).and_then(|python| match python {
-                Some(python) => Ok(python),
-                None => prepare_managed_runtime(&server, &progress),
-            }) {
-                Ok(python) => python,
-                Err(err) => {
-                    return LocalModelServerStatus {
-                        state: LocalModelServerState::Error,
-                        backend: backend.to_string(),
-                        model: model.to_string(),
-                        detail: format_error_chain(&err),
-                        server_path: Some(server.path),
-                        endpoint: None,
-                    };
-                }
-            };
-
-            let model_path = match prepare_local_embedding_model_snapshot(
-                &server.runtime_dir,
-                &profile,
-                &progress,
-            ) {
-                Ok(path) => path,
-                Err(err) => {
-                    return LocalModelServerStatus {
-                        state: LocalModelServerState::Error,
-                        backend: backend.to_string(),
-                        model: model.to_string(),
-                        detail: format!(
-                            "failed to prepare model files:\n{}",
-                            format_error_chain(&err)
-                        ),
-                        server_path: Some(server.path),
-                        endpoint: None,
-                    };
-                }
-            };
-
-            start_model_server(&server, &python, backend, model, model_path.as_deref())
-        }
-        Err(err) => LocalModelServerStatus {
-            state: LocalModelServerState::Error,
-            backend: backend.to_string(),
-            model: model.to_string(),
-            detail: err.to_string(),
-            server_path: None,
-            endpoint: None,
-        },
-    }
-}
-
-pub(crate) fn inspect_prepared_runtime_status(
-    server: &BundledModelServer,
-    backend: &str,
-    model: &str,
-) -> LocalModelServerStatus {
-    let python = venv_python_path(&server.venv_dir);
-    match prepared_runtime_python(server) {
-        Ok(Some(_)) => LocalModelServerStatus {
-            state: LocalModelServerState::PreparedButStopped,
-            backend: backend.to_string(),
-            model: model.to_string(),
-            detail: "managed Python runtime is prepared; model server is not running".to_string(),
-            server_path: Some(server.path.clone()),
-            endpoint: None,
-        },
-        Ok(None) => LocalModelServerStatus {
-            state: LocalModelServerState::SetupRequired,
-            backend: backend.to_string(),
-            model: model.to_string(),
-            detail: if python.is_file() {
-                "managed Python venv exists but dependencies are not installed".to_string()
-            } else {
-                format!("venv python missing at {}", python.display())
-            },
-            server_path: Some(server.path.clone()),
-            endpoint: None,
-        },
-        Err(err) => LocalModelServerStatus {
-            state: LocalModelServerState::Error,
-            backend: backend.to_string(),
-            model: model.to_string(),
-            detail: format!("failed to inspect managed Python runtime: {err}"),
-            server_path: Some(server.path.clone()),
-            endpoint: None,
-        },
-    }
-}
-
-pub(crate) fn prepared_runtime_python(server: &BundledModelServer) -> Result<Option<PathBuf>> {
-    let python = venv_python_path(&server.venv_dir);
-    if !python.is_file() {
-        return Ok(None);
-    }
-    let requirements = selected_requirements_file(server)?;
-    if !requirements_marker_matches(
-        &requirements_marker_path(&server.runtime_dir),
-        &requirements.sha256,
-    )? {
-        return Ok(None);
-    }
-    Ok(Some(python))
-}
-
-pub(crate) fn prepare_managed_runtime(
-    server: &BundledModelServer,
-    progress: &Option<LocalProgressReporter>,
-) -> Result<PathBuf> {
-    let python =
-        ensure_managed_venv(server, progress).context("failed to create managed Python venv")?;
-    if let Err(err) = ensure_model_server_dependencies(server, &python, progress)
-        .context("failed to install model server dependencies")
-    {
-        if let Err(cleanup_err) = cleanup_failed_venv(server) {
-            return Err(anyhow!(
-                "{err}; also failed to remove incomplete managed venv {}: {cleanup_err}",
-                server.venv_dir.display()
-            ));
-        }
-        return Err(anyhow!(
-            "{err}; removed incomplete managed venv {}",
-            server.venv_dir.display()
-        ));
-    }
-    Ok(python)
-}
-
-pub(crate) fn cleanup_failed_venv(server: &BundledModelServer) -> Result<()> {
-    ensure_runtime_dir_inside_home(&server.runtime_dir, &server.venv_dir)?;
-    if server.venv_dir.exists() {
-        fs::remove_dir_all(&server.venv_dir)
-            .with_context(|| format!("remove failed managed venv {}", server.venv_dir.display()))?;
-    }
-    Ok(())
-}
-
-pub(crate) fn reusable_server_status(
-    server: &BundledModelServer,
-    backend: &str,
-    model: &str,
-) -> Option<LocalModelServerStatus> {
-    let metadata = read_server_metadata(&metadata_path(&server.runtime_dir)).ok()??;
-    if metadata.component_sha256 != server.sha256 || metadata.profile != embedding_profile_id() {
-        return None;
-    }
-    let endpoint = endpoint_url(&metadata.host, metadata.port);
-    let health = probe_health(&metadata.host, metadata.port).ok()?;
-    if !health_identity_matches(&health, backend, model) {
-        return None;
-    }
-    if !health_model_ready(&health, backend) {
-        if let Some(error_message) = health_preparation_error(&health, backend) {
-            return Some(LocalModelServerStatus {
-                state: LocalModelServerState::Error,
-                backend: backend.to_string(),
-                model: model.to_string(),
-                detail: format!(
-                    "model server running at {endpoint}; model preparation failed: {error_message}"
-                ),
-                server_path: Some(server.path.clone()),
-                endpoint: Some(endpoint),
-            });
-        }
-        return Some(LocalModelServerStatus {
-            state: LocalModelServerState::PreparingModel,
-            backend: backend.to_string(),
-            model: model.to_string(),
-            detail: format!("model server running at {endpoint}; model is still loading"),
-            server_path: Some(server.path.clone()),
-            endpoint: Some(endpoint),
-        });
-    }
-    Some(LocalModelServerStatus {
-        state: LocalModelServerState::Ready,
-        backend: backend.to_string(),
-        model: model.to_string(),
-        detail: format!("reusing model server at {endpoint}"),
-        server_path: Some(server.path.clone()),
-        endpoint: Some(endpoint),
-    })
-}
-
-pub(crate) fn start_model_server(
-    server: &BundledModelServer,
-    python: &Path,
-    backend: &str,
-    model: &str,
-    model_path: Option<&Path>,
-) -> LocalModelServerStatus {
-    let host = DEFAULT_MODEL_SERVER_HOST;
-    let port = DEFAULT_MODEL_SERVER_PORT;
-    let endpoint = endpoint_url(host, port);
-    let model_cache_dir = default_local_model_cache_dir();
-    let stderr_log_path = server.runtime_dir.join("model-server-stderr.log");
-    let stderr_file = std::fs::File::create(&stderr_log_path)
-        .unwrap_or_else(|_| std::fs::File::create("/dev/null").unwrap());
-    match Command::new(python)
-        .arg(&server.path)
-        .env("RARA_MODEL_SERVER_HOST", host)
-        .env("RARA_MODEL_SERVER_PORT", port.to_string())
-        .env("RARA_MODEL_CACHE_DIR", &model_cache_dir)
-        .stdin(Stdio::null())
-        .stdout(Stdio::null())
-        .stderr(Stdio::from(stderr_file))
-        .spawn()
-    {
-        Ok(mut child) => {
-            let metadata = ModelServerMetadata {
-                pid: child.id(),
-                host: host.to_string(),
-                port,
-                component_sha256: server.sha256.clone(),
-                profile: embedding_profile_id().to_string(),
-                started_at: unix_timestamp_secs(),
-            };
-            if let Err(err) = write_server_metadata(&metadata_path(&server.runtime_dir), &metadata)
-            {
-                return LocalModelServerStatus {
-                    state: LocalModelServerState::Error,
-                    backend: backend.to_string(),
-                    model: model.to_string(),
-                    detail: format!("failed to write model server metadata: {err}"),
-                    server_path: Some(server.path.clone()),
-                    endpoint: Some(endpoint),
-                };
-            }
-
-            let mut health_ever_passed = false;
-            let mut last_prepare_error: Option<anyhow::Error> = None;
-            for _ in 0..STARTUP_HEALTH_ATTEMPTS {
-                if let Ok(health) = probe_health(host, port) {
-                    health_ever_passed = true;
-                    if health_identity_matches(&health, backend, model) {
-                        match prepare_model(host, port, backend, model_path) {
-                            Ok(state) => {
-                                let (state_tag, detail) = if state == "ready" {
-                                    (
-                                        LocalModelServerState::Ready,
-                                        format!(
-                                            "started model server and prepared model at {endpoint}"
-                                        ),
-                                    )
-                                } else {
-                                    (
-                                        LocalModelServerState::PreparingModel,
-                                        format!(
-                                            "started model server at {endpoint}; model is still loading"
-                                        ),
-                                    )
-                                };
-                                return LocalModelServerStatus {
-                                    state: state_tag,
-                                    backend: backend.to_string(),
-                                    model: model.to_string(),
-                                    detail,
-                                    server_path: Some(server.path.clone()),
-                                    endpoint: Some(endpoint),
-                                };
-                            }
-                            Err(err) => {
-                                last_prepare_error = Some(err);
-                            }
-                        }
-                    }
-                }
-                std::thread::sleep(STARTUP_HEALTH_DELAY);
-                if let Ok(Some(status)) = child.try_wait() {
-                    return LocalModelServerStatus {
-                        state: LocalModelServerState::Error,
-                        backend: backend.to_string(),
-                        model: model.to_string(),
-                        detail: format!(
-                            "model server process exited with {status} during startup (stderr log: {})",
-                            stderr_log_path.display()
-                        ),
-                        server_path: Some(server.path.clone()),
-                        endpoint: Some(endpoint),
-                    };
-                }
-            }
-            if let Some(err) = last_prepare_error {
-                return LocalModelServerStatus {
-                    state: LocalModelServerState::Error,
-                    backend: backend.to_string(),
-                    model: model.to_string(),
-                    detail: format!(
-                        "failed to prepare model at {endpoint} after {} health checks: {err} (stderr log: {})",
-                        STARTUP_HEALTH_ATTEMPTS,
-                        stderr_log_path.display()
-                    ),
-                    server_path: Some(server.path.clone()),
-                    endpoint: Some(endpoint),
-                };
-            }
-
-            let stderr_hint = format!(" (stderr log: {})", stderr_log_path.display());
-            match prepare_model(host, port, backend, model_path) {
-                Ok(state) => {
-                    health_ever_passed = true;
-                    if state == "ready" {
-                        return LocalModelServerStatus {
-                            state: LocalModelServerState::Ready,
-                            backend: backend.to_string(),
-                            model: model.to_string(),
-                            detail: format!(
-                                "started model server and prepared model at {endpoint}"
-                            ),
-                            server_path: Some(server.path.clone()),
-                            endpoint: Some(endpoint),
-                        };
-                    }
-                    if let Ok(health) = probe_health(host, port) {
-                        if health_identity_matches(&health, backend, model) {
-                            return LocalModelServerStatus {
-                                state: LocalModelServerState::PreparingModel,
-                                backend: backend.to_string(),
-                                model: model.to_string(),
-                                detail: format!(
-                                    "started model server at {endpoint}; model is still loading"
-                                ),
-                                server_path: Some(server.path.clone()),
-                                endpoint: Some(endpoint),
-                            };
-                        }
-                    }
-                }
-                Err(err) => {
-                    last_prepare_error = Some(err);
-                }
-            }
-
-            LocalModelServerStatus {
-                state: LocalModelServerState::Starting,
-                backend: backend.to_string(),
-                model: model.to_string(),
-                detail: if health_ever_passed {
-                    format!(
-                        "started model server process; waiting for health at {endpoint}{stderr_hint}"
-                    )
-                } else {
-                    format!(
-                        "started model server process; health never reached at {endpoint}{stderr_hint}"
-                    )
-                },
-                server_path: Some(server.path.clone()),
-                endpoint: Some(endpoint),
-            }
-        }
-        Err(err) => LocalModelServerStatus {
-            state: LocalModelServerState::Error,
-            backend: backend.to_string(),
-            model: model.to_string(),
-            detail: format!("failed to start model server: {err}"),
-            server_path: Some(server.path.clone()),
-            endpoint: Some(endpoint),
-        },
-    }
-}
-
-pub(crate) fn default_embedding_profile() -> LocalEmbeddingModelProfile {
-    if cfg!(all(target_os = "macos", target_arch = "aarch64")) {
-        LocalEmbeddingModelProfile {
-            backend: "mlx_qwen3",
-            model: MLX_QWEN3_MODEL_ID,
-            revision: MODEL_REVISION,
-            snapshot_preparation: SnapshotPreparation::RustManaged {
-                required_files: SnapshotRequiredFiles::MlxQwen3,
-            },
-        }
-    } else {
-        LocalEmbeddingModelProfile {
-            backend: "fastembed_bge_m3",
-            model: FASTEMBED_BGE_M3_MODEL_ID,
-            revision: MODEL_REVISION,
-            snapshot_preparation: SnapshotPreparation::PythonManaged,
-        }
-    }
-}
-
-pub(crate) fn default_embedding_backend() -> (&'static str, &'static str) {
-    let profile = default_embedding_profile();
-    (profile.backend, profile.model)
-}
-
-pub(crate) fn embedding_profile_id() -> &'static str {
-    "qwen3-embedding-0.6b"
-}
-
-pub(crate) fn venv_python_path(venv_dir: &Path) -> PathBuf {
-    if cfg!(windows) {
-        venv_dir.join("Scripts").join("python.exe")
-    } else {
-        venv_dir.join("bin").join("python")
-    }
-}
-
-/// Find a Python >= 3.10 on the system.
-///
-/// Respects `RARA_PYTHON` if set; otherwise probes known executable names in
-/// descending version order and picks the first one whose reported version is
-/// at least 3.10.
-pub(crate) fn ensure_uv_installed() -> Result<()> {
-    match Command::new("uv")
-        .arg("--version")
-        .stdin(Stdio::null())
-        .output()
-    {
-        Ok(output) if output.status.success() => Ok(()),
-        _ => bail!(
-            "uv is not installed; install it:\n  https://docs.astral.sh/uv/getting-started/installation/"
-        ),
-    }
-}
-
-pub(crate) fn ensure_managed_venv(
-    server: &BundledModelServer,
-    progress: &Option<LocalProgressReporter>,
-) -> Result<PathBuf> {
-    let python = venv_python_path(&server.venv_dir);
-    if python.is_file() {
-        return Ok(python);
-    }
-    if server.venv_dir.exists() {
-        fs::remove_dir_all(&server.venv_dir)
-            .with_context(|| format!("remove stale venv {}", server.venv_dir.display()))?;
-    }
-    report_progress(progress, "Embedding · creating Python venv with uv");
-    ensure_uv_installed()?;
-    let output = Command::new("uv")
-        .arg("venv")
-        .arg("--python")
-        .arg("3.14")
-        .arg("--seed")
-        .arg(&server.venv_dir)
-        .stdin(Stdio::null())
-        .output()
-        .with_context(|| "run uv venv --python 3.14 --seed")?;
-    ensure_command_success(output, "create managed Python venv with uv")?;
-    if !python.is_file() {
-        bail!("venv python was not created at {}", python.display());
-    }
-    Ok(python)
-}
-
-pub(crate) fn ensure_model_server_dependencies(
-    server: &BundledModelServer,
-    python: &Path,
-    progress: &Option<LocalProgressReporter>,
-) -> Result<()> {
-    let requirements = selected_requirements_file(server)?;
-    let marker_path = requirements_marker_path(&server.runtime_dir);
-    if requirements_marker_matches(&marker_path, &requirements.sha256)? {
-        return Ok(());
-    }
-
-    report_progress(
-        progress,
-        "Embedding · installing model server dependencies with uv",
-    );
-    let output = Command::new("uv")
-        .arg("pip")
-        .arg("install")
-        .arg("--python")
-        .arg(python)
-        .arg("-r")
-        .arg(&requirements.path)
-        .stdin(Stdio::null())
-        .output()
-        .with_context(|| "run uv pip install")?;
-    ensure_command_success(output, "install model server dependencies with uv")?;
-
-    let marker = serde_json::json!({
-        "requirements_sha256": requirements.sha256,
-        "requirements_path": requirements.path,
-    });
-    write_file_atomically(&marker_path, serde_json::to_vec_pretty(&marker)?.as_slice())?;
-    Ok(())
-}
-
-pub(crate) fn selected_requirements_file(
-    server: &BundledModelServer,
-) -> Result<&BundledModelServerFile> {
-    let name = if cfg!(all(target_os = "macos", target_arch = "aarch64")) {
-        REQUIREMENTS_MACOS_ARM64_NAME
-    } else {
-        REQUIREMENTS_PORTABLE_NAME
-    };
-    server
-        .requirements
-        .iter()
-        .find(|file| file.path.ends_with(name))
-        .ok_or_else(|| anyhow!("missing bundled requirements file {name}"))
-}
-
-pub(crate) fn requirements_marker_path(runtime_dir: &Path) -> PathBuf {
-    runtime_dir.join(REQUIREMENTS_MARKER_NAME)
-}
-
-pub(crate) fn requirements_marker_matches(path: &Path, expected_hash: &str) -> Result<bool> {
-    let content = match fs::read_to_string(path) {
-        Ok(content) => content,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(false),
-        Err(err) => return Err(err).with_context(|| format!("read {}", path.display())),
-    };
-    let value: Value =
-        serde_json::from_str(&content).with_context(|| format!("parse {}", path.display()))?;
-    Ok(value
-        .get("requirements_sha256")
-        .and_then(Value::as_str)
-        .is_some_and(|hash| hash == expected_hash))
-}
-
-pub(crate) fn prepare_local_embedding_model_snapshot(
-    runtime_dir: &Path,
-    profile: &LocalEmbeddingModelProfile,
-    progress: &Option<LocalProgressReporter>,
-) -> Result<Option<PathBuf>> {
-    let SnapshotPreparation::RustManaged { required_files } = profile.snapshot_preparation else {
-        return Ok(None);
-    };
-    let model = profile.model;
-
-    let cache_dir = default_local_model_cache_dir();
-    let marker_path = model_snapshot_marker_path(runtime_dir);
-    if let Some(marker) =
-        read_matching_model_snapshot_marker(&marker_path, model, profile.revision)?
-    {
-        if cached_snapshot_under_cache(&marker.snapshot_path, &cache_dir)?
-            && snapshot_has_all_files(&marker.snapshot_path, &marker.files)
-        {
-            report_progress(
-                progress,
-                format!(
-                    "Model · already available at {}",
-                    marker.snapshot_path.display()
-                ),
-            );
-            return Ok(Some(marker.snapshot_path));
-        }
-    }
-
-    let repo = Repo::with_revision(
-        model.to_string(),
-        RepoType::Model,
-        profile.revision.to_string(),
-    );
-    let cache = Cache::new(cache_dir.clone());
-    let cache_repo = cache.repo(repo.clone());
-    report_progress(
-        progress,
-        format!("Model · checking local snapshot for {model}"),
-    );
-    if let Some((snapshot_path, files)) =
-        local_cached_model_snapshot(&cache_dir, &repo, required_files, profile.revision)?
-    {
-        write_model_snapshot_marker(
-            &marker_path,
-            model,
-            profile.revision,
-            &snapshot_path,
-            &files,
-        )?;
-        report_progress(
-            progress,
-            format!("Model · already available at {}", snapshot_path.display()),
-        );
-        return Ok(Some(snapshot_path));
-    }
-
-    let mut builder = ApiBuilder::from_cache(cache)
-        .with_progress(false)
-        .with_retries(3);
-    if let Ok(endpoint) = std::env::var("HF_ENDPOINT") {
-        builder = builder.with_endpoint(endpoint);
-    }
-    if let Some(token) = std::env::var("HF_TOKEN")
-        .ok()
-        .filter(|value| !value.is_empty())
-    {
-        builder = builder.with_token(Some(token));
-    }
-    let api = builder.build().context("build Hugging Face API client")?;
-    let api_repo = api.repo(repo);
-    report_progress(
-        progress,
-        format!("Model · resolving model metadata for {model}"),
-    );
-    let info = api_repo
-        .info()
-        .context("resolve model repository metadata")?;
-    let files: Vec<String> = info
-        .siblings
-        .into_iter()
-        .map(|sibling| sibling.rfilename)
-        .filter(|name| !name.ends_with('/'))
-        .collect();
-    if files.is_empty() {
-        bail!("model repository has no downloadable files");
-    }
-
-    let snapshot_path = cache_repo.pointer_path(&info.sha);
-    if snapshot_has_all_files(&snapshot_path, &files) {
-        write_model_snapshot_marker(
-            &marker_path,
-            model,
-            profile.revision,
-            &snapshot_path,
-            &files,
-        )?;
-        report_progress(
-            progress,
-            format!("Model · already available at {}", snapshot_path.display()),
-        );
-        return Ok(Some(snapshot_path));
-    }
-
-    report_progress(
-        progress,
-        format!("Model · downloading {} file(s)", files.len()),
-    );
-    for filename in &files {
-        let target = snapshot_path.join(filename);
-        if target.exists() {
-            report_progress(progress, format!("Model · cached {filename}"));
-            continue;
-        }
-        report_progress(progress, format!("Model · downloading {filename}"));
-        api_repo
-            .download_with_progress(
-                filename,
-                TuiDownloadProgress::new(filename.clone(), progress.clone()),
-            )
-            .with_context(|| format!("download model file {filename}"))?;
-    }
-
-    if !snapshot_has_all_files(&snapshot_path, &files) {
-        bail!("model snapshot is incomplete after download");
-    }
-    write_model_snapshot_marker(
-        &marker_path,
-        model,
-        profile.revision,
-        &snapshot_path,
-        &files,
-    )?;
-    report_progress(
-        progress,
-        format!("Model · ready at {}", snapshot_path.display()),
-    );
-    Ok(Some(snapshot_path))
-}
-
-pub(crate) fn snapshot_has_all_files(snapshot_path: &Path, files: &[String]) -> bool {
-    files
-        .iter()
-        .all(|filename| snapshot_path.join(filename).exists())
-}
-
-pub(crate) fn local_cached_model_snapshot(
-    cache_dir: &Path,
-    repo: &Repo,
-    required_files: SnapshotRequiredFiles,
-    revision: &str,
-) -> Result<Option<(PathBuf, Vec<String>)>> {
-    let repo_dir = cache_dir.join(repo.folder_name());
-    let ref_path = repo_dir.join("refs").join(revision);
-    let commit_hash = match fs::read_to_string(&ref_path) {
-        Ok(hash) => hash.trim().to_string(),
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(None),
-        Err(err) => return Err(err).with_context(|| format!("read {}", ref_path.display())),
-    };
-    if commit_hash.is_empty() {
-        return Ok(None);
-    }
-    let snapshot_path = repo_dir.join("snapshots").join(commit_hash);
-    if !cached_snapshot_under_cache(&snapshot_path, cache_dir)? {
-        return Ok(None);
-    }
-    let files = collect_snapshot_files(&snapshot_path)?;
-    if files.is_empty() || !snapshot_has_minimum_model_files(required_files, &files) {
-        return Ok(None);
-    }
-    Ok(Some((snapshot_path, files)))
-}
-
-pub(crate) fn collect_snapshot_files(snapshot_path: &Path) -> Result<Vec<String>> {
-    let mut files = Vec::new();
-    collect_snapshot_files_inner(snapshot_path, snapshot_path, &mut files)?;
-    files.sort();
-    Ok(files)
-}
-
-pub(crate) fn collect_snapshot_files_inner(
-    root: &Path,
-    current: &Path,
-    files: &mut Vec<String>,
-) -> Result<()> {
-    let entries = match fs::read_dir(current) {
-        Ok(entries) => entries,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(()),
-        Err(err) => return Err(err).with_context(|| format!("read {}", current.display())),
-    };
-    for entry in entries {
-        let entry = entry.with_context(|| format!("read entry in {}", current.display()))?;
-        let path = entry.path();
-        let file_type = entry
-            .file_type()
-            .with_context(|| format!("stat {}", path.display()))?;
-        if file_type.is_dir() {
-            collect_snapshot_files_inner(root, &path, files)?;
-        } else if file_type.is_file() || file_type.is_symlink() {
-            let relative = path
-                .strip_prefix(root)
-                .with_context(|| format!("strip {}", root.display()))?;
-            files.push(relative_path_string(relative));
-        }
-    }
-    Ok(())
-}
-
-pub(crate) fn relative_path_string(path: &Path) -> String {
-    path.components()
-        .filter_map(|component| match component {
-            std::path::Component::Normal(part) => Some(part.to_string_lossy()),
-            _ => None,
-        })
-        .collect::<Vec<_>>()
-        .join("/")
-}
-
-pub(crate) fn snapshot_has_minimum_model_files(
-    required_files: SnapshotRequiredFiles,
-    files: &[String],
-) -> bool {
-    match required_files {
-        SnapshotRequiredFiles::MlxQwen3 => {
-            let has_config = files.iter().any(|file| file == "config.json");
-            let has_tokenizer = files
-                .iter()
-                .any(|file| file == "tokenizer.json" || file == "tokenizer.model");
-            let has_weights = files.iter().any(|file| file.ends_with(".safetensors"));
-            has_config && has_tokenizer && has_weights
-        }
-    }
-}
-
-pub(crate) fn model_snapshot_marker_path(runtime_dir: &Path) -> PathBuf {
-    runtime_dir.join(MODEL_SNAPSHOT_MARKER_NAME)
-}
-
-pub(crate) fn read_matching_model_snapshot_marker(
-    path: &Path,
-    expected_model: &str,
-    expected_revision: &str,
-) -> Result<Option<ModelSnapshotMarker>> {
-    let content = match fs::read_to_string(path) {
-        Ok(content) => content,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(None),
-        Err(err) => return Err(err).with_context(|| format!("read {}", path.display())),
-    };
-    let marker: ModelSnapshotMarker =
-        serde_json::from_str(&content).with_context(|| format!("parse {}", path.display()))?;
-    if marker.model != expected_model || marker.revision != expected_revision {
-        return Ok(None);
-    }
-    if marker.files.is_empty() {
-        return Ok(None);
-    }
-    Ok(Some(marker))
-}
-
-pub(crate) fn write_model_snapshot_marker(
-    path: &Path,
-    model: &str,
-    revision: &str,
-    snapshot_path: &Path,
-    files: &[String],
-) -> Result<()> {
-    let marker = ModelSnapshotMarker {
-        model: model.to_string(),
-        revision: revision.to_string(),
-        snapshot_path: snapshot_path.to_path_buf(),
-        files: files.to_vec(),
-    };
-    write_file_atomically(path, serde_json::to_vec_pretty(&marker)?.as_slice())
-}
-
-pub(crate) fn cached_snapshot_under_cache(snapshot_path: &Path, cache_dir: &Path) -> Result<bool> {
-    let snapshot = match fs::canonicalize(snapshot_path) {
-        Ok(path) => path,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(false),
-        Err(err) => {
-            return Err(err).with_context(|| format!("resolve {}", snapshot_path.display()));
-        }
-    };
-    let cache = match fs::canonicalize(cache_dir) {
-        Ok(path) => path,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(false),
-        Err(err) => return Err(err).with_context(|| format!("resolve {}", cache_dir.display())),
-    };
-    Ok(snapshot == cache || snapshot.starts_with(cache))
-}
-
-pub(crate) fn report_progress(
-    progress: &Option<LocalProgressReporter>,
-    message: impl Into<String>,
-) {
-    if let Some(callback) = progress {
-        callback(message.into());
-    }
-}
-
-pub(crate) struct TuiDownloadProgress {
-    filename: String,
-    progress: Option<LocalProgressReporter>,
-    total: usize,
-    current: usize,
-    last_percent: Option<usize>,
-}
-
-impl TuiDownloadProgress {
-    fn new(filename: String, progress: Option<LocalProgressReporter>) -> Self {
-        Self {
-            filename,
-            progress,
-            total: 0,
-            current: 0,
-            last_percent: None,
-        }
-    }
-
-    fn emit(&mut self, force: bool) {
-        let percent = if self.total == 0 {
-            0
-        } else {
-            self.current.saturating_mul(100) / self.total
-        };
-        if !force && self.last_percent == Some(percent) {
-            return;
-        }
-        self.last_percent = Some(percent);
-        report_progress(
-            &self.progress,
-            format!(
-                "Model · {} · {}% ({}/{})",
-                self.filename,
-                percent,
-                format_bytes(self.current),
-                format_bytes(self.total)
-            ),
-        );
-    }
-}
-
-impl HfProgress for TuiDownloadProgress {
-    fn init(&mut self, size: usize, filename: &str) {
-        self.total = size;
-        self.current = 0;
-        self.filename = filename.to_string();
-        self.emit(true);
-    }
-
-    fn update(&mut self, size: usize) {
-        self.current = self.current.saturating_add(size);
-        self.emit(false);
-    }
-
-    fn finish(&mut self) {
-        self.current = self.total;
-        self.emit(true);
-    }
-}
-
-pub(crate) fn format_bytes(bytes: usize) -> String {
-    const KIB: f64 = 1024.0;
-    const MIB: f64 = KIB * 1024.0;
-    const GIB: f64 = MIB * 1024.0;
-    let value = bytes as f64;
-    if value >= GIB {
-        format!("{:.1}GiB", value / GIB)
-    } else if value >= MIB {
-        format!("{:.1}MiB", value / MIB)
-    } else if value >= KIB {
-        format!("{:.1}KiB", value / KIB)
-    } else {
-        format!("{bytes}B")
-    }
-}
-
-pub(crate) fn ensure_model_server_requirements(
-    runtime_dir: &Path,
-) -> Result<Vec<BundledModelServerFile>> {
-    let requirements_dir = runtime_dir.join("requirements");
-    fs::create_dir_all(&requirements_dir)
-        .with_context(|| format!("create {}", requirements_dir.display()))?;
-
-    [
-        BundledFile {
-            name: REQUIREMENTS_MACOS_ARM64_NAME,
-            content: REQUIREMENTS_MACOS_ARM64,
-        },
-        BundledFile {
-            name: REQUIREMENTS_PORTABLE_NAME,
-            content: REQUIREMENTS_PORTABLE,
-        },
-    ]
-    .into_iter()
-    .map(|file| {
-        let path = requirements_dir.join(file.name);
-        let expected_hash = sha256_hex(file.content);
-        if !existing_file_matches(&path, &expected_hash)? {
-            write_file_atomically(&path, file.content)?;
-        }
-        Ok(BundledModelServerFile {
-            path,
-            sha256: expected_hash,
-        })
-    })
-    .collect()
-}
-
-impl StartupLock {
-    fn try_acquire(path: &Path) -> Result<Self> {
-        if let Some(parent) = path.parent() {
-            fs::create_dir_all(parent)
-                .with_context(|| format!("create lock directory {}", parent.display()))?;
-        }
-        let file = fs::OpenOptions::new()
-            .read(true)
-            .write(true)
-            .create(true)
-            .truncate(false)
-            .open(path)
-            .with_context(|| format!("open lock file {}", path.display()))?;
-        file.try_lock_exclusive()
-            .with_context(|| format!("lock file {}", path.display()))?;
-        Ok(Self { _file: file })
-    }
-}
-
-impl Drop for StartupLock {
-    fn drop(&mut self) {
-        let _ = self._file.unlock();
-    }
-}
-
-pub(crate) fn metadata_path(runtime_dir: &Path) -> PathBuf {
-    runtime_dir.join(SERVER_METADATA_NAME)
-}
-
-pub(crate) fn startup_lock_path(runtime_dir: &Path) -> PathBuf {
-    runtime_dir.join(STARTUP_LOCK_NAME)
-}
-
-pub(crate) fn endpoint_url(host: &str, port: u16) -> String {
-    format!("http://{host}:{port}")
-}
-
-pub(crate) fn read_server_metadata(path: &Path) -> Result<Option<ModelServerMetadata>> {
-    let content = match fs::read_to_string(path) {
-        Ok(content) => content,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(None),
-        Err(err) => return Err(err).with_context(|| format!("read {}", path.display())),
-    };
-    let metadata = serde_json::from_str(&content)
-        .with_context(|| format!("parse model server metadata {}", path.display()))?;
-    Ok(Some(metadata))
-}
-
-pub(crate) fn write_server_metadata(path: &Path, metadata: &ModelServerMetadata) -> Result<()> {
-    let content = serde_json::to_vec_pretty(metadata)?;
-    write_file_atomically(path, &content)
-}
-
-pub(crate) fn probe_health(host: &str, port: u16) -> Result<Value> {
-    model_server_http_client(HEALTH_TIMEOUT)?
-        .get(model_server_url(host, port, "/health")?)
-        .header(reqwest::header::ACCEPT, "application/json")
-        .send()
-        .context("send model server health request")?
-        .error_for_status()
-        .context("model server health returned non-success status")?
-        .json()
-        .context("parse model server health response")
-}
-
-pub(crate) fn prepare_model(
-    host: &str,
-    port: u16,
-    backend: &str,
-    model_path: Option<&Path>,
-) -> Result<String> {
-    let mut body = serde_json::json!({ "backend": backend });
-    if let Some(model_path) = model_path {
-        body["model_path"] = Value::String(model_path.display().to_string());
-    }
-    let body = body.to_string();
-    let url = format!("http://{host}:{port}/models/prepare");
-    let response = post_json(host, port, "/models/prepare", &body, PREPARE_TIMEOUT)
-        .with_context(|| format!("call POST {url}"))?;
-    if response.get("ok").and_then(Value::as_bool) != Some(true) {
-        bail!("model prepare endpoint returned non-ok response");
-    }
-    if response.get("backend").and_then(Value::as_str) != Some(backend) {
-        bail!("model prepare endpoint returned mismatched backend");
-    }
-    let state = response
-        .get("state")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown")
-        .to_string();
-    if state != "ready" && state != "loading" {
-        bail!("model prepare endpoint returned unexpected state: {state}");
-    }
-    Ok(state)
-}
-
-pub(crate) fn post_json(
-    host: &str,
-    port: u16,
-    path: &str,
-    body: &str,
-    timeout: Duration,
-) -> Result<Value> {
-    let response = model_server_http_client(timeout)?
-        .post(model_server_url(host, port, path)?)
-        .header(reqwest::header::ACCEPT, "application/json")
-        .header(reqwest::header::CONTENT_TYPE, "application/json")
-        .body(body.to_string())
-        .send()
-        .context("send model server POST request")?;
-    if !response.status().is_success() {
-        let status = response.status();
-        let error = response
-            .text()
-            .unwrap_or_else(|_| "missing response body".to_string());
-        bail!("model server POST returned non-success status {status}: {error}");
-    }
-    response.json().context("parse model server POST response")
-}
-
-pub(crate) fn model_server_http_client(timeout: Duration) -> Result<reqwest::blocking::Client> {
-    reqwest::blocking::Client::builder()
-        .timeout(timeout)
-        .no_proxy()
-        .build()
-        .context("build model server HTTP client")
-}
-
-pub(crate) fn model_server_url(host: &str, port: u16, path: &str) -> Result<reqwest::Url> {
-    reqwest::Url::parse(&format!("http://{host}:{port}{path}"))
-        .with_context(|| format!("build model server URL for {host}:{port}{path}"))
-}
-
-pub(crate) fn health_identity_matches(health: &Value, backend: &str, model: &str) -> bool {
-    if health.get("ok").and_then(Value::as_bool) != Some(true) {
-        return false;
-    }
-    if health
-        .get("default_embedding_backend")
-        .and_then(Value::as_str)
-        != Some(backend)
-    {
-        return false;
-    }
-    health
-        .get("embeddings")
-        .and_then(|value| value.get(backend))
-        .and_then(|value| value.get("model"))
-        .and_then(Value::as_str)
-        .is_none_or(|reported| reported == model)
-}
-
-pub(crate) fn health_model_ready(health: &Value, backend: &str) -> bool {
-    health
-        .get("embeddings")
-        .and_then(|value| value.get(backend))
-        .and_then(|value| value.get("loaded"))
-        .and_then(Value::as_bool)
-        == Some(true)
-}
-
-pub(crate) fn health_preparation_error<'a>(health: &'a Value, backend: &str) -> Option<&'a str> {
-    let prep = health
-        .get("preparation")
-        .and_then(|value| value.get(backend))?;
-    if prep.get("state").and_then(Value::as_str) == Some("error") {
-        prep.get("message").and_then(Value::as_str)
-    } else {
-        None
-    }
-}
-
-pub(crate) fn ensure_command_success(output: Output, context: &str) -> Result<()> {
-    if output.status.success() {
-        return Ok(());
-    }
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    bail!(
-        "{context} failed with status {}: {}{}{}",
-        output.status,
-        stderr.trim(),
-        if stdout.trim().is_empty() { "" } else { "\n" },
-        stdout.trim()
-    );
-}
-
-pub(crate) fn unix_timestamp_secs() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_secs()
-}
-
-pub(crate) fn ensure_runtime_dir_inside_home(rara_home: &Path, runtime_dir: &Path) -> Result<()> {
-    let home = fs::canonicalize(rara_home)
-        .with_context(|| format!("canonicalize {}", rara_home.display()))?;
-    let runtime = fs::canonicalize(runtime_dir)
-        .with_context(|| format!("canonicalize {}", runtime_dir.display()))?;
-    if !runtime.starts_with(&home) {
-        bail!(
-            "refusing to install model server outside RARA home: {}",
-            runtime.display()
-        );
-    }
-    Ok(())
-}
-
-pub(crate) fn existing_file_matches(path: &Path, expected_hash: &str) -> Result<bool> {
-    let metadata = match fs::symlink_metadata(path) {
-        Ok(metadata) => metadata,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(false),
-        Err(err) => return Err(err).with_context(|| format!("stat {}", path.display())),
-    };
-    if metadata.file_type().is_symlink() {
-        bail!(
-            "refusing to overwrite symlinked model server: {}",
-            path.display()
-        );
-    }
-    if !metadata.is_file() {
-        bail!(
-            "refusing to overwrite non-file model server path: {}",
-            path.display()
-        );
-    }
-    let content = fs::read(path).with_context(|| format!("read {}", path.display()))?;
-    Ok(sha256_hex(&content) == expected_hash)
-}
-
-pub(crate) fn write_file_atomically(path: &Path, content: &[u8]) -> Result<()> {
-    if let Ok(metadata) = fs::symlink_metadata(path) {
-        if metadata.file_type().is_symlink() {
-            bail!(
-                "refusing to overwrite symlinked model server: {}",
-                path.display()
-            );
-        }
-    }
-
-    let file_name = path
-        .file_name()
-        .and_then(|name| name.to_str())
-        .ok_or_else(|| anyhow!("invalid model server path: {}", path.display()))?;
-    let write_id = ATOMIC_WRITE_COUNTER.fetch_add(1, Ordering::Relaxed);
-    let tmp_path = path.with_file_name(format!(
-        ".{file_name}.{}.{}.tmp",
-        std::process::id(),
-        write_id
-    ));
-    let mut file = fs::OpenOptions::new()
-        .write(true)
-        .create_new(true)
-        .open(&tmp_path)
-        .with_context(|| format!("create {}", tmp_path.display()))?;
-    file.write_all(content)
-        .with_context(|| format!("write {}", tmp_path.display()))?;
-    file.sync_all()
-        .with_context(|| format!("sync {}", tmp_path.display()))?;
-    drop(file);
-
-    set_private_file_permissions(&tmp_path)?;
-    fs::rename(&tmp_path, path)
-        .with_context(|| format!("rename {} to {}", tmp_path.display(), path.display()))?;
-    Ok(())
-}
-
-#[cfg(unix)]
-pub(crate) fn set_private_file_permissions(path: &Path) -> Result<()> {
-    use std::os::unix::fs::PermissionsExt;
-
-    let permissions = fs::Permissions::from_mode(0o600);
-    fs::set_permissions(path, permissions)
-        .with_context(|| format!("chmod 0600 {}", path.display()))?;
-    Ok(())
-}
-
-#[cfg(not(unix))]
-pub(crate) fn set_private_file_permissions(_path: &Path) -> Result<()> {
-    Ok(())
-}
+include!("_types.rs");
+include!("_backend.rs");
+include!("_status.rs");
+include!("_server.rs");
+include!("_prepare.rs");
+include!("_model.rs");
+include!("_progress.rs");
+include!("_util.rs");
 
 pub(crate) fn sha256_hex(content: &[u8]) -> String {
     let digest = Sha256::digest(content);
